% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_coords.R
\name{find_coords}
\alias{find_coords}
\title{Find coordinates of a tensor}
\usage{
find_coords(cellcenter, tipUnit, d, scale = 1)
}
\arguments{
\item{cellcenter}{x, y, z coordinates}

\item{tipUnit}{x, y, z coordinates}

\item{d}{norm of the vector}

\item{scale}{shorten or lengthen the line for visualisation}
}
\description{
Find coordinates of a tensor
}
