% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh2ply.R
\name{mesh2ply}
\alias{mesh2ply}
\title{Write a ply format for MorphoGraphX from a mesh3D R object}
\usage{
mesh2ply(mesh = mesh, filename = "my_ply.ply",
  label_it_for_mesh = mesh$label_it)
}
\arguments{
\item{mesh}{A 3D (triangular) mesh object.}

\item{filename}{The name we want to give to the created .ply file.}

\item{label_it_for_mesh}{Triangle color.}
}
\value{
a .ply file readable by MorphoGraphX
}
\description{
Write a ply format for MorphoGraphX from a mesh3D R object
}
\keyword{mesh3D,}
\keyword{ply}
