% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d2ply_clust.R
\name{mesh3D2ply_clust}
\alias{mesh3D2ply_clust}
\title{Write a ply format for MorphoGraphX from a mesh3D R object}
\usage{
mesh3D2ply_clust(mesh = mesh, filename = "my_ply.ply",
  label_it_for_mesh = mesh$label_it)
}
\arguments{
\item{mesh}{a 3D (triangular) mesh object}

\item{filename}{the filename we want to give to the .ply file}

\item{label_it_for_mesh}{triangle color}
}
\value{
a .ply file readable by MorphoGraphX
mesh3D2ply_clust()
}
\description{
original code: 2016-09-11_mesh3D_to_ply_clust
}
\keyword{mesh3D,}
\keyword{ply}
