% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgx2r-package.R
\docType{package}
\name{mgx2r-package}
\alias{mgx2r-package}
\alias{mgx2r}
\title{mgx2r-package}
\description{
Import mesh and cell graph files from MorphoGraphX software as tidy dataframes.
Propose visualization solutions for mesh and cell graph files from MorphoGraphX with Plotly and Shiny.
}
\author{
Marion <marion.louveaux@gmail.com>
}
