% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cellGraph_function.R
\name{modified_read.cellGraph}
\alias{modified_read.cellGraph}
\title{A reading .ply Function}
\usage{
modified_read.cellGraph(cellGraphfile, header_max = 30)
}
\arguments{
\item{file}{.ply file path. The .ply file stores information relative to a triangular mesh created with MorphoGraphX.}

\item{ShowSpecimen}{Default to TRUE. Display the newly created mesh3D with the chosen color (defined in MatCol).}

\item{addNormals}{Default to TRUE}

\item{MatCol}{Default to "signal". Expect either "signal" (fluorescent signal projected on the mesh), "parent" (cell label) or "label" (cell face label). To fill Material$color of mesh 3D.}

\item{my_colors}{colors for displaying the mesh in an hexadecimal format}
}
\value{
Object of type mesh3D
modified_read.ply()
}
\description{
This function is a modified version of read.ply function from geomorph R package to specifically read MGX ply files.
}
\keyword{.ply,}
\keyword{read}
