% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyMesh.R
\name{plotlyMesh}
\alias{plotlyMesh}
\title{A layout for plotly graphs}
\usage{
plotlyMesh(meshExample, meshColors = NULL, meshCellcenter = NULL)
}
\arguments{
\item{meshExample}{mesh loaded with the modified_read.ply function}

\item{meshColors}{to display colored on the mesh}

\item{meshCellcenter}{to display the center of (biological) cells on the mesh}
}
\value{
plotly graph
plotlyMesh()
}
\description{
This function provides a basic layout for plotly graphs used to display MGX meshes.
}
\keyword{plotly}
