% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyMeshall.R
\name{plotlyMesh_all}
\alias{plotlyMesh_all}
\title{A layout for plotly graphs with a time slider}
\usage{
plotlyMesh_all(meshExample, graphExample, meshColors, display,
  defaultColor = list("#CCCCFF", 0.2), heatmapParam = "GeometryArea")
}
\arguments{
\item{meshExample}{One or several meshes.}

\item{graphExample}{One or several cell graphs.}

\item{meshColors}{Colors for each mesh.}

\item{display}{Display type. 'labels', 'heatmap' or 'none'.}

\item{defaultColor}{Mesh color if 'none' display type is chosen (HEX color and alpha value between 0 and 1).}

\item{heatmapParam}{Variable (from cell graph) to display if 'heatmap' display type is selected.}
}
\value{
plotly graph
}
\description{
This function provides a basic layout for plotly graphs with a time sliders.
It can be used to display MGX meshes of a time serie.
}
