% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mgxCellGraph.R
\name{read_mgxCellGraph}
\alias{read_mgxCellGraph}
\title{A reading .ply Function}
\usage{
read_mgxCellGraph(fileCellGraph, header_max = 30)
}
\arguments{
\item{fileCellGraph}{.ply Cell Graph file.}

\item{header_max}{Number of lines expected in header. Must be equal or greater to the actual number of lines in the header. Default to 30.}
}
\value{
Object of type mesh3D
}
\description{
This function is a modified version of read.ply function from geomorph R package to specifically read MGX cell graph ply files.
}
\keyword{.ply,}
\keyword{read}
