% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cellcontour.R
\name{show_cellcontour}
\alias{show_cellcontour}
\title{Add cell contour visualisation to a plotly graph of type mesh3d}
\usage{
show_cellcontour(p, mesh)
}
\arguments{
\item{p}{a plotly graph created with plotlyMesh()}

\item{mesh}{the mesh used to create this plotly graph}
}
\description{
Add cell contour visualisation to a plotly graph of type mesh3d
}
