% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tensors.R
\name{show_tensors}
\alias{show_tensors}
\title{Add tensor visualisation to a plotly graph of type mesh3d}
\usage{
show_tensors(p, cellGraph, tensor_name, scale = 1)
}
\arguments{
\item{p}{A plot created with plotlyMesh() function.}

\item{cellGraph}{A cell graph created with read_cellGraph().}

\item{tensor_name}{The name of the tensor variable from cell graph to visualise.}

\item{scale}{Shorten or lengthen the line for visualisation.}
}
\value{
a plotly graph
}
\description{
Add tensor visualisation to a plotly graph of type mesh3d
}
