% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ply_function.R
\name{modified_read.ply}
\alias{modified_read.ply}
\title{A reading .ply Function}
\usage{
modified_read.ply(file, ShowSpecimen = TRUE, addNormals = TRUE,
  MatCol = 1, header_max = 30, my_colors = c("#800000", "#FF0000",
  "#808000", "#FFFF00", "#008000", "#00FF00", "#008080", "#00FFFF", "#000080",
  "#0000FF", "#800080", "#FF00FF"))
}
\arguments{
\item{file}{.ply file path. The .ply file stores information relative to a triangular mesh created with MorphoGraphX.}

\item{ShowSpecimen}{Default to TRUE. Display the newly created mesh3D with the chosen color (defined in MatCol).}

\item{addNormals}{Default to TRUE.}

\item{MatCol}{Integer. Default to 1, e.g. first element the header which is not x, y or z. To fill Material$color of mesh 3D.}

\item{header_max}{Number of lines expected in header. Must be equal or greater to the actual number of lines in the header. Default to 30.}

\item{my_colors}{Colors for displaying the mesh in an hexadecimal format.}
}
\value{
Object of type mesh3D
}
\description{
This function is a modified version of read.ply function from geomorph R package to specifically read MGX mesh ply files.
}
\keyword{.ply,}
\keyword{read}
