% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucharHelp.R
\name{ucharHelp}
\alias{ucharHelp}
\title{Identify and sort uchar in the properties of vertices and faces}
\usage{
ucharHelp(ppty, file = file, toSkip = toSkip, Nlines = Nlines)
}
\arguments{
\item{ppty}{Lines containing the properties info in the header of the .ply.}

\item{file}{Ply file.}

\item{toSkip}{To skip when reading the file (either headerend or headerend+nvertices if reading the face ppty).}

\item{Nlines}{Number of face or vertices.}
}
\description{
Identify and sort uchar in the properties of vertices and faces
}
